add env var support
